Implement Combatant base class
